Uses `for..in` to generate object.values
As the comment states, this doesn't cover all the edge cases that the
true `Object.values` takes into account. **However,** because we are
using this on a plain object which we are creating, we can be confident
that the edge cases don't apply in this instant.

For me, this justifies its usage.